+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 * Add support for `OAUTH_ACCESS_REVOKED` in `WebhookNotification`s
 * Add support for dispute search by `customerId`, `disbursementDate`, and `effectiveDate`
 * Add `payerId` accessor in `PayPalAccount`
+* Update `jackson-jr` to v2.9.5 to fix an incompatibility with Java 9
 
 ## 2.77.0
 * Fix possible NullPointerException in UsBankAccount